Harvested service   validation

+++ N2K_BackboneBackEnd/Helpers/AzureBlobHandler.cs
@@ -5,13 +5,11 @@ using System.Net.Http.Headers;
 
 namespace N2K_BackboneBackEnd.Helpers
 {
-    public class AzureBlobHandler : IAttachedFileHandler
+    public class AzureBlobHandler : AttachedFileHandler, IAttachedFileHandler
     {
 
-        private readonly AttachedFilesConfig _attachedFilesConfig;
-        public AzureBlobHandler(AttachedFilesConfig attachedFilesConfig)
+        public AzureBlobHandler(AttachedFilesConfig attachedFilesConfig) : base(attachedFilesConfig)
         {
-            _attachedFilesConfig = attachedFilesConfig;
         }
 
 
@@ -33,7 +31,8 @@ namespace N2K_BackboneBackEnd.Helpers
             var folderName = _attachedFilesConfig.JustificationFolder;
             var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), folderName);
             var remoteUrl = "";
-            List <String> uploadedFiles = new List<string>();
+            List<String> uploadedFiles = new List<string>();
+            var invalidFile = false;
 
             foreach (var f in files.Files)
             {
@@ -41,34 +40,71 @@ namespace N2K_BackboneBackEnd.Helpers
                 var fullPath = Path.Combine(pathToSave, fileName);
                 using (var stream = new FileStream(fullPath, FileMode.Create))
                 {
-                    f.CopyTo(stream);
+                    await f.CopyToAsync(stream);
                 }
-
-                //if the file is compressed (extract all the content)
-                if (fullPath.EndsWith("zip"))
+                if (!CheckExtensions(fileName) || invalidFile == true)
+                {
+                    invalidFile = true;
+                    File.Delete(fullPath);
+                    break;
+                }
+                if (fileName.EndsWith("zip"))
                 {
                     using (ZipArchive archive = ZipFile.OpenRead(fullPath))
                     {
                         archive.ExtractToDirectory(pathToSave);
                         foreach (ZipArchiveEntry entry in archive.Entries)
                         {
-                            BlobClient blobClient1 = ConnectToAzureBlob().GetBlobClient(entry.Name);
-                            await blobClient1.UploadAsync(Path.Combine(pathToSave, entry.Name), true);
-                            remoteUrl = _attachedFilesConfig.PublicFilesUrl + (!_attachedFilesConfig.PublicFilesUrl.EndsWith("/") ? "/" : "");
-                            uploadedFiles.Add(string.Format("{0}{1}/{2}", remoteUrl, folderName, entry.Name));
-
-                            File.Delete(Path.Combine(pathToSave,entry.Name ));
+                            if (!CheckExtensions(entry.Name))
+                            {
+                                invalidFile = true;
+                                File.Delete(Path.Combine(pathToSave, entry.Name));
+                                break;
+                            }
+                            File.Delete(Path.Combine(pathToSave, entry.Name));
                         }
                     }
                 }
-                else
-                {
-                    BlobClient blobClient = ConnectToAzureBlob().GetBlobClient(fileName);
-                    await blobClient.UploadAsync(fullPath, true);
-                    remoteUrl = _attachedFilesConfig.PublicFilesUrl + (!_attachedFilesConfig.PublicFilesUrl.EndsWith("/") ? "/" : "");
-                    uploadedFiles.Add(string.Format("{0}{1}/{2}", remoteUrl, folderName, fileName));
-                }
                 File.Delete(fullPath);
+            }
+
+            if (!invalidFile)
+            {
+                foreach (var f in files.Files)
+                {
+                    var fileName = ContentDispositionHeaderValue.Parse(f.ContentDisposition).FileName.Trim('"');
+                    var fullPath = Path.Combine(pathToSave, fileName);
+                    using (var stream = new FileStream(fullPath, FileMode.Create))
+                    {
+                        f.CopyTo(stream);
+                    }
+
+                    //if the file is compressed (extract all the content)
+                    if (fullPath.EndsWith("zip"))
+                    {
+                        using (ZipArchive archive = ZipFile.OpenRead(fullPath))
+                        {
+                            archive.ExtractToDirectory(pathToSave);
+                            foreach (ZipArchiveEntry entry in archive.Entries)
+                            {
+                                BlobClient blobClient1 = ConnectToAzureBlob().GetBlobClient(entry.Name);
+                                await blobClient1.UploadAsync(Path.Combine(pathToSave, entry.Name), true);
+                                remoteUrl = _attachedFilesConfig.PublicFilesUrl + (!_attachedFilesConfig.PublicFilesUrl.EndsWith("/") ? "/" : "");
+                                uploadedFiles.Add(string.Format("{0}{1}/{2}", remoteUrl, folderName, entry.Name));
+
+                                File.Delete(Path.Combine(pathToSave, entry.Name));
+                            }
+                        }
+                    }
+                    else
+                    {
+                        BlobClient blobClient = ConnectToAzureBlob().GetBlobClient(fileName);
+                        await blobClient.UploadAsync(fullPath, true);
+                        remoteUrl = _attachedFilesConfig.PublicFilesUrl + (!_attachedFilesConfig.PublicFilesUrl.EndsWith("/") ? "/" : "");
+                        uploadedFiles.Add(string.Format("{0}{1}/{2}", remoteUrl, folderName, fileName));
+                    }
+                    File.Delete(fullPath);
+                }
             }
             return uploadedFiles;
